requirements_docker.txt
# Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# NLP/AI
sentence-transformers==2.3.1
spacy==3.7.2
transformers==4.36.0
huggingface-hub==0.20.0
torch==2.1.0+cpu
torchvision==0.16.0+cpu

# Document parsing
pdfplumber==0.10.4
python-docx==1.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0

# Frontend
streamlit==1.31.0
plotly==5.18.0

# Development
black==24.1.0
ruff==0.1.0

pandas==2.1.4
requests==2.31.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl


--extra-index-url https://download.pytorch.org/whl/cpu